Extract pytorch iterator tests to a new file
