+++ qa/TL0_python-self-test-core/test_body.sh
@@ -40,7 +40,7 @@ test_type_annotations() {
 
 test_pytorch() {
     ${python_invoke_test} --attr '!slow,pytorch' test_dali_variable_batch_size.py
-    ${python_new_invoke_test} -A '!slow,pytorch' -s checkpointing
+    ${python_new_invoke_test} -A '!slow' checkpointing.test_dali_checkpointing_fw_iterators.TestPytorch
     ${python_new_invoke_test} -A 'pytorch' -s type_annotations
 }
